generator: carry a JSON number exactly from the schema to the generated source (#215, #216, #220) (#230)
A JSON number was read into a float64 at the schema boundary, and every
consumer downstream then worked from the rounding rather than from the
number. The three issues are that one fact surfacing wherever the rounding
happened to land.

- #215: the emitted integer decode fell back to float64 and guarded with
  `_f <= 9.223372036854776e+18`, which is 2^63 -- one past MaxInt64.
  `int64(_f)` on 2^63 is undefined and yields -2^63, so 9223372036854775808
  was accepted, sign-flipped, and round-tripped wrong with no error anywhere.
- #216: `{"type":"integer","const":9223372036854775807}` emitted
  9.223372036854776e+18 into an int64 constant declaration, which does not
  compile. Below that threshold the same rounding compiled and gave the wrong
  verdict instead: a const of 9007199254740993 rejected its own value and
  accepted 9007199254740992.
- #220: under --big-int the value was kept as a big.Int and the *bound* was
  re-emitted as its shortest decimal, which big.Float.SetString then read
  literally: 2^62 became 4611686018427388000 and refused a value equal to its
  own minimum.

The numeric keywords are now schema.Number, a literal kept as written that
answers Float64, Int64 and Rat for itself; const, enum and default keep their
numbers because the schema decoder uses UseNumber. Precision comes from the
literal and range still comes from float64, so a keyword float64 cannot hold
is refused exactly where it was before.

Downstream, each consumer decides what the number has to become rather than
being handed one reading of it: GoNumberLiteral writes the constant, the
bigint templates get the literal for SetString, and a bound on an int64
instance is compared in int64 -- 2^53 is where consecutive integers stop
having distinct float64s, and above it the conversion of the *value* was
losing the comparison on its own.

The generated integer decode reads the literal as decimal digits
(jsonIntegerFromLiteral) instead of parsing a float64 and converting back,
which is the only way to tell 9223372036854775807 from 9223372036854775808.

Two things are deliberately unchanged, and both are the same limitation:
the JSON form a raw enum and a const check compare against is still built
through float64, on both sides, because the emitted check decodes the
instance into `any` and marshals it back. Writing the literal on one side
only would turn an enum of 1.0 into a rejection of the document 1.

tests/numeric_precision_test.go generates, compiles and runs the boundary
matrix -- MaxInt64 and its neighbours, 2^63, MinInt64, 2^53 -- under both
configs, checking the decode, the verdict and the byte-exact round trip.

+++ testdata/golden/advanced/complex_tuple.go
@@ -171,12 +171,12 @@ func (e EventRecordItem2) Validate() error {
 		}
 	}
 	if e._jsonKeys["code"] {
-		if e.Code != nil && float64(*e.Code) < 100 {
+		if e.Code != nil && *e.Code < 100 {
 			return fmt.Errorf("code: value %v is less than minimum 100", *e.Code)
 		}
 	}
 	if e._jsonKeys["code"] {
-		if e.Code != nil && float64(*e.Code) > 599 {
+		if e.Code != nil && *e.Code > 599 {
 			return fmt.Errorf("code: value %v exceeds maximum 599", *e.Code)
 		}
 	}
